+++ packages/payload/src/collections/operations/countVersions.ts
@@ -1,12 +1,13 @@
 import type { AccessResult } from '../../config/types.js'
+import type { CollectionSlug } from '../../index.js'
 import type { PayloadRequest, Where } from '../../types/index.js'
 import type { Collection } from '../config/types.js'
 
 import executeAccess from '../../auth/executeAccess.js'
 import { combineQueries } from '../../database/combineQueries.js'
-import { validateQueryPaths } from '../../database/queryValidation/validateQueryPaths.js'
-import { buildVersionCollectionFields, type CollectionSlug } from '../../index.js'
+import { validateQueryPaths } from '../../database/queryValidation/validateSearchParams.js'
 import { killTransaction } from '../../utilities/killTransaction.js'
+import { buildVersionCollectionFields } from '../../versions/buildCollectionFields.js'
 import { buildAfterOperation } from './utils.js'
 
 export type Arguments = {
