+++ packages/payload/src/collections/operations/update.ts
@@ -16,8 +16,8 @@ import type {
 import { ensureUsernameOrEmail } from '../../auth/ensureUsernameOrEmail.js'
 import executeAccess from '../../auth/executeAccess.js'
 import { combineQueries } from '../../database/combineQueries.js'
-import { validateQueryPaths } from '../../database/queryValidation/validateQueryPaths.js'
-import { APIError } from '../../errors/index.js'
+import { validateQueryPaths } from '../../database/queryValidation/validateSearchParams.js'
+import { APIError } from '../../errors/APIError.js'
 import { afterChange } from '../../fields/hooks/afterChange/index.js'
 import { afterRead } from '../../fields/hooks/afterRead/index.js'
 import { beforeChange } from '../../fields/hooks/beforeChange/index.js'
