+++ packages/payload/src/database/queryValidation/validateSearchParams.ts
@@ -1,14 +1,17 @@
 import type { SanitizedCollectionConfig } from '../../collections/config/types.js'
-import type { Field } from '../../fields/config/types.js'
+import type { Field, FieldAffectingData } from '../../fields/config/types.js'
 import type { SanitizedGlobalConfig } from '../../globals/config/types.js'
-import type { PayloadRequest } from '../../types/index.js'
+import type { Operator, PayloadRequest, Where, WhereField } from '../../types/index.js'
 import type { EntityPolicies, PathToQuery } from './types.js'
 
+import { QueryError } from '../../errors/QueryError.js'
 import { fieldAffectsData, fieldIsVirtual } from '../../fields/config/types.js'
+import { validOperators } from '../../types/constants.js'
+import { deepCopyObject } from '../../utilities/deepCopyObject.js'
+import { flattenTopLevelFields } from '../../utilities/flattenTopLevelFields.js'
 import { getEntityPolicies } from '../../utilities/getEntityPolicies.js'
 import isolateObjectProperty from '../../utilities/isolateObjectProperty.js'
 import { getLocalizedPaths } from '../getLocalizedPaths.js'
-import { validateQueryPaths } from './validateQueryPaths.js'
 
 type Args = {
   collectionConfig?: SanitizedCollectionConfig
@@ -192,4 +195,85 @@ export async function validateSearchParam({
     }),
   )
   await Promise.all(promises)
+}
+
+type ValidateQueryPathsArgs = {
+  errors?: { path: string }[]
+  overrideAccess: boolean
+  policies?: EntityPolicies
+  req: PayloadRequest
+  versionFields?: Field[]
+  where: Where
+} & (
+  | {
+      collectionConfig: SanitizedCollectionConfig
+      globalConfig?: never | undefined
+    }
+  | {
+      collectionConfig?: never | undefined
+      globalConfig: SanitizedGlobalConfig
+    }
+)
+
+const flattenWhere = (query: Where): WhereField[] =>
+  Object.entries(query).reduce((flattenedConstraints, [key, val]) => {
+    if ((key === 'and' || key === 'or') && Array.isArray(val)) {
+      const subWhereConstraints: Where[] = val.reduce((acc, subVal) => {
+        const subWhere = flattenWhere(subVal)
+        return [...acc, ...subWhere]
+      }, [])
+      return [...flattenedConstraints, ...subWhereConstraints]
+    }
+
+    return [...flattenedConstraints, { [key]: val }]
+  }, [])
+
+export async function validateQueryPaths({
+  collectionConfig,
+  errors = [],
+  globalConfig,
+  overrideAccess,
+  policies = {
+    collections: {},
+    globals: {},
+  },
+  req,
+  versionFields,
+  where,
+}: ValidateQueryPathsArgs): Promise<void> {
+  const fields = flattenTopLevelFields(
+    versionFields || (globalConfig || collectionConfig).fields,
+  ) as FieldAffectingData[]
+  if (typeof where === 'object') {
+    const whereFields = flattenWhere(where)
+    // We need to determine if the whereKey is an AND, OR, or a schema path
+    const promises = []
+    void whereFields.map((constraint) => {
+      void Object.keys(constraint).map((path) => {
+        void Object.entries(constraint[path]).map(([operator, val]) => {
+          if (validOperators.includes(operator as Operator)) {
+            promises.push(
+              validateSearchParam({
+                collectionConfig: deepCopyObject(collectionConfig),
+                errors,
+                fields: fields as Field[],
+                globalConfig: deepCopyObject(globalConfig),
+                operator,
+                overrideAccess,
+                path,
+                policies,
+                req,
+                val,
+                versionFields,
+              }),
+            )
+          }
+        })
+      })
+    })
+    await Promise.all(promises)
+    if (errors.length > 0) {
+      throw new QueryError(errors)
+    }
+  }
 }
