+++ packages/payload/src/index.ts
@@ -1006,8 +1006,10 @@ export { readMigrationFiles } from './database/migrations/readMigrationFiles.js'
 export { writeMigrationIndex } from './database/migrations/writeMigrationIndex.js'
 export type * from './database/queryValidation/types.js'
 export type { EntityPolicies, PathToQuery } from './database/queryValidation/types.js'
-export { validateQueryPaths } from './database/queryValidation/validateQueryPaths.js'
-export { validateSearchParam } from './database/queryValidation/validateSearchParams.js'
+export {
+  validateQueryPaths,
+  validateSearchParam,
+} from './database/queryValidation/validateSearchParams.js'
 export type {
   BaseDatabaseAdapter,
   BeginTransaction,
@@ -1200,8 +1202,10 @@ export type {
 
 export { getDefaultValue } from './fields/getDefaultValue.js'
 export { traverseFields as afterChangeTraverseFields } from './fields/hooks/afterChange/traverseFields.js'
-export { promise as afterReadPromise } from './fields/hooks/afterRead/promise.js'
-export { traverseFields as afterReadTraverseFields } from './fields/hooks/afterRead/traverseFields.js'
+export {
+  promise as afterReadPromise,
+  traverseFields as afterReadTraverseFields,
+} from './fields/hooks/afterRead/traverseFields.js'
 export { traverseFields as beforeChangeTraverseFields } from './fields/hooks/beforeChange/traverseFields.js'
 export { traverseFields as beforeValidateTraverseFields } from './fields/hooks/beforeValidate/traverseFields.js'
 export { default as sortableFieldTypes } from './fields/sortableFieldTypes.js'
@@ -1324,7 +1328,7 @@ export {
   pathExistsAndIsAccessible,
   pathExistsAndIsAccessibleSync,
 } from './utilities/findUp.js'
-export { default as flattenTopLevelFields } from './utilities/flattenTopLevelFields.js'
+export { flattenTopLevelFields } from './utilities/flattenTopLevelFields.js'
 export { formatErrors } from './utilities/formatErrors.js'
 export { formatLabels, formatNames, toWords } from './utilities/formatLabels.js'
 export { getCollectionIDFieldTypes } from './utilities/getCollectionIDFieldTypes.js'
